Add files via upload
a python program that returns the top X recurring words and their counts from a text document by utilizing a hash map consisting of an array of linked lists. To execute, place a text document in the same fold/level as word_count.py and change the number of X recurring words as desired.